+++ CHANGELOG.md
@@ -51,6 +51,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - HID Device: set Report Mode as default
 - HID Host: support large HID Descriptors
 - AVDTP: limit L2CAP MTU to largest BR/EDR packet, fixes A2DP Sink role for some Samsung TV
+- AVDTP: fix Get Stream Configuration
 - AVRCP: reject incoming second l2cap connection
 - AVRCP: generate command for avrcp_controller_get_element_attributes dynamically
 - SPP Server: fix SPP version in SDP record